requirements.txt

langchain
langchain-core
langchain-community
langchain-text-splitters
pypdf
sentence-transformers
chromadb
numpy
python-dotenv
langchain-google-genai
langgraph
unstructured
python-docx
langchain-mcp-adapters
fastapi
uvicorn
mcp